Added documentation for employee

+++ employee/views.py
@@ -6,8 +6,14 @@ from .forms import *
 from .models import Employee
 
 
+# create an employee using a supplied form
 @login_required
 def create_employee(request, title, header, button, form_class):
+    # make sure the request is made by a Manager
+    try:
+        Manager.objects.get(pk=request.user.id)
+    except:
+        return redirect('home:home')
     # check if the request is POST
     if request.method == 'POST':
         # keep a copy of the signup fill info
@@ -16,11 +22,11 @@ def create_employee(request, title, header, button, form_class):
         if form.is_valid():
             # create a user object without submitting it to the database
             user = form.save(commit=False)
-            # make sure the user object cannot log in unless activated
+            # make sure the user is active
             user.is_active = True
             # commit the user to the database
             user.save()
-
+            # render the success page
             return render(request, '../templates/success.html')
     else:
         # if the form wasn't filled before, make a new empty form.
@@ -29,7 +35,7 @@ def create_employee(request, title, header, button, form_class):
     return render(request, '../templates/render_form.html', {'form': form, 'title':title, 'header': header, 'button':button})
 
 
-# defines the signup process
+# pass in a form and information to be rendered
 @login_required
 def create_teller(request):
     title = 'Teller'
@@ -38,6 +44,7 @@ def create_teller(request):
     return create_employee(request, title, header, button, TellerForm)
 
 
+# pass in a form and information to be rendered
 @login_required
 def create_advisor(request):
     title = 'Advisor'
@@ -80,9 +87,13 @@ def restore_employee(request, id):
 
 
 ########### Manager Management
+
+
+# pass in a form and information to be rendered
 @login_required
 def mgr_create(request):
-    if not Manager.objects.get(pk=request.user.id):
+    # if the user is not admin they cannot make a manager
+    if not request.user.is_superuser:
         return redirect('home:home')
     title = 'Manager'
     header = 'Create Manager'
@@ -90,32 +101,53 @@ def mgr_create(request):
     return create_employee(request, title, header, button, ManagerForm)
 
 
+# Render the homepage for the manager management with search and a list of managers
 @login_required
 def mgr_home(request):
+    # if the user is not admin they cannot be on the manager page
+    if not request.user.is_superuser:
+        return redirect('home:home')
+    # get all managers
     managers = Manager.objects.all().order_by('id')
+    # display the home
     return render(request, 'Mgr_management/home.html', {'managers': managers})
 
 
+# display the results of a search
 @login_required
 def mgr_search(request):
+    # if the user is not admin they cannot search a manager
+    if not request.user.is_superuser:
+        return redirect('home:home')
     searched = request.GET.get('results', '')
-    # retrieve all matching posts
+    # check if there was something searched
     if searched:
-        # Perform your search logic here, for example:
+        # check data for matches
         managers = Manager.objects.filter(Q(f_name__contains=searched) | Q(l_name__contains=searched) | Q(id__contains=searched) | Q(dept__name__contains=searched))
     else:
         managers = Manager.objects.none()
+    # display results
     return render(request, 'Mgr_management/search.html', {'searched': searched, 'managers': managers})
 
 
+# Display the information of a manager
 @login_required
 def mgr_info(request, mgr_id):
+    # if the user is not admin they cannot see a manager
+    if not request.user.is_superuser:
+        return redirect('home:home')
+    # check if manager exists
     mgr = get_object_or_404(Manager, id=mgr_id)
     return render(request, 'Mgr_management/employee_info.html', {'manager': mgr})
 
 
+# edit manager fields
 @login_required
 def mgr_edit(request, mgr_id):
+    # if the user is not admin they cannot edit a manager
+    if not request.user.is_superuser:
+        return redirect('home:home')
+    # populate the render fields
     title = 'Update'
     header = 'Update Department'
     button = 'Submit'
@@ -124,11 +156,10 @@ def mgr_edit(request, mgr_id):
     if request.method == 'POST':
         # Create a form instance with the submitted data
         form = ManagerForm(request.POST, request.FILES, instance=mgr)
-
         if form.is_valid():
             # Save the updated model instance
             form.save()
-
+            # redirect to previous url
             previous_url = request.META.get('HTTP_REFERER')
             if previous_url:
                 return redirect(previous_url, )
@@ -138,16 +169,22 @@ def mgr_edit(request, mgr_id):
     else:
         # Create a form instance with the data from the model instance to be updated
         form = ManagerForm(instance=mgr)
-
     # Render the update form template with the form and model instance
     return render(request, '../templates/render_form.html',
                   {'form': form, 'title': title, 'header': header, 'button': button})
 
 
+# delete the manager
 @login_required
 def mgr_delete(request, mgr_id):
+    # if the user is not admin they cannot delete a manager
+    if not request.user.is_superuser:
+        return redirect('home:home')
+    # check if the manager exists
     mgr = get_object_or_404(Manager, id=mgr_id)
+    # delete the manager
     mgr.delete()
+    # redirect to previous page
     previous_url = request.META.get('HTTP_REFERER')
     if previous_url:
         return redirect(previous_url, )
